knowledge_graph/graph.py
# knowledge_graph/graph.py
import networkx as nx

class Graph:
    def __init__(self):
        self.graph = nx.Graph()

    def add_node(self, node_id, node_type):
        self.graph.add_node(node_id, type=node_type)

    def add_edge(self, source_id, target_id, relationship):
        self.graph.add_edge(source_id, target_id, relationship=relationship)

    def get_node(self, node_id):
        return self.graph.nodes[node_id]

    def query(self, query):
        # Placeholder for query logic
        return []